chore: move heuristics check to another folder
