+++ src/macaron/malware_analyzer/pypi_heuristics/analysis_result.py
