+++ src/macaron/malware_analyzer/pypi_heuristics/base_analyzer.py
@@ -5,9 +5,9 @@
 
 from abc import abstractmethod
 
+from macaron.malware_analyzer.pypi_heuristics.analysis_result import HeuristicResult
+from macaron.malware_analyzer.pypi_heuristics.heuristics import HEURISTIC
 from macaron.slsa_analyzer.package_registry.pypi_registry import PyPIRegistry
-from macaron.slsa_analyzer.pypi_heuristics.analysis_result import HeuristicResult
-from macaron.slsa_analyzer.pypi_heuristics.heuristics import HEURISTIC
 
 
 class BaseHeuristicAnalyzer:
