+++ src/macaron/malware_analyzer/pypi_heuristics/heuristics.py
