+++ src/macaron/malware_analyzer/pypi_heuristics/metadata/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (c) 2022 - 2024, Oracle and/or its affiliates. All rights reserved.
+# Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl/.
