+++ src/macaron/malware_analyzer/pypi_heuristics/metadata/one_release.py
@@ -4,10 +4,10 @@
 
 """Analyzer checks the packages contain one release."""
 
+from macaron.malware_analyzer.pypi_heuristics.analysis_result import HeuristicResult
+from macaron.malware_analyzer.pypi_heuristics.base_analyzer import BaseHeuristicAnalyzer
+from macaron.malware_analyzer.pypi_heuristics.heuristics import HEURISTIC
 from macaron.slsa_analyzer.package_registry.pypi_registry import PyPIRegistry
-from macaron.slsa_analyzer.pypi_heuristics.analysis_result import HeuristicResult
-from macaron.slsa_analyzer.pypi_heuristics.base_analyzer import BaseHeuristicAnalyzer
-from macaron.slsa_analyzer.pypi_heuristics.heuristics import HEURISTIC
 
 
 class OneReleaseAnalyzer(BaseHeuristicAnalyzer):
