+++ src/macaron/malware_analyzer/pypi_heuristics/metadata/unchanged_release.py
@@ -6,10 +6,10 @@ import logging
 from collections import Counter
 
 from macaron.json_tools import json_extract
+from macaron.malware_analyzer.pypi_heuristics.analysis_result import HeuristicResult
+from macaron.malware_analyzer.pypi_heuristics.base_analyzer import BaseHeuristicAnalyzer
+from macaron.malware_analyzer.pypi_heuristics.heuristics import HEURISTIC
 from macaron.slsa_analyzer.package_registry.pypi_registry import PyPIRegistry
-from macaron.slsa_analyzer.pypi_heuristics.analysis_result import HeuristicResult
-from macaron.slsa_analyzer.pypi_heuristics.base_analyzer import BaseHeuristicAnalyzer
-from macaron.slsa_analyzer.pypi_heuristics.heuristics import HEURISTIC
 
 logger: logging.Logger = logging.getLogger(__name__)
