+++ src/macaron/malware_analyzer/pypi_heuristics/sourcecode/suspicious_setup.py
@@ -13,10 +13,10 @@ import zipfile
 
 import requests
 
+from macaron.malware_analyzer.pypi_heuristics.analysis_result import HeuristicResult
+from macaron.malware_analyzer.pypi_heuristics.base_analyzer import BaseHeuristicAnalyzer
+from macaron.malware_analyzer.pypi_heuristics.heuristics import HEURISTIC
 from macaron.slsa_analyzer.package_registry.pypi_registry import PyPIRegistry
-from macaron.slsa_analyzer.pypi_heuristics.analysis_result import HeuristicResult
-from macaron.slsa_analyzer.pypi_heuristics.base_analyzer import BaseHeuristicAnalyzer
-from macaron.slsa_analyzer.pypi_heuristics.heuristics import HEURISTIC
 
 logger: logging.Logger = logging.getLogger(__name__)
