+++ src/macaron/slsa_analyzer/analyzer.py
@@ -31,6 +31,9 @@ from macaron.errors import (
     PURLNotFoundError,
     RepoCheckOutError,
 )
+from macaron.malware_analyzer.checks import (  # pylint: disable=unused-import # noqa: F401
+    detect_malicious_metadata_check,
+)
 from macaron.output_reporter.reporter import FileReporter
 from macaron.output_reporter.results import Record, Report, SCMStatus
 from macaron.repo_finder import repo_finder
